requirements.txt
langchain
streamlit
python-dotenv
cassandra-driver
openai
cassio
